+++ app/controllers/api/v1/api_controller.rb
@@ -1,0 +1,5 @@
+module Api::V1
+    class ApiController < ApplicationController
+      
+    end
+end
