+++ app/controllers/application_controller.rb
@@ -1,2 +1,2 @@
-class ApplicationController < ActionController::Base
+class ApplicationController < ActionController::API
 end
