+++ config/routes.rb
@@ -1,6 +1,6 @@
 Rails.application.routes.draw do
-  namespace 'api' do
-    namespace 'v1' do
+  namespace 'api', constraints: { format: 'json' } do
+    namespace :v1 do
       resources :campgrounds
     end
   end
